requirements.txt
Django>=5.2
gunicorn
whitenoise[brotli]
dj-database-url
psycopg2-binary
django-ckeditor
Pillow